refactor: quack button 1으로 롤백
