+++ feature/profile/src/main/kotlin/team/duckie/app/android/feature/profile/screen/section/ButtonSection.kt
@@ -7,21 +7,20 @@
 
 package team.duckie.app.android.feature.profile.screen.section
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import team.duckie.app.android.feature.profile.R
-import team.duckie.quackquack.material.QuackColor
-import team.duckie.quackquack.material.QuackTypography
-import team.duckie.quackquack.material.quackClickable
-import team.duckie.quackquack.ui.QuackText
+import team.duckie.quackquack.ui.border.QuackBorder
+import team.duckie.quackquack.ui.color.QuackColor
+import team.duckie.quackquack.ui.component.QuackSurface
+import team.duckie.quackquack.ui.component.internal.QuackText
+import team.duckie.quackquack.ui.textstyle.QuackTextStyle
 
 @Composable
 internal fun FollowSection(
@@ -47,39 +46,39 @@ fun EditButton(
     enabled: Boolean = false,
     plainTextColor: QuackColor = QuackColor.Gray1,
 ) {
-    QuackText(
-        modifier = modifier
-            .clip(RoundedCornerShape(size = 8.dp))
-            .quackClickable(onClick = onClick)
-            .background(
-                if (enabled) {
-                    QuackColor.White.toBrush()
-                } else {
-                    QuackColor.Gray4.toBrush()
-                },
-            )
-            .border(
+    QuackSurface(
+        modifier = modifier.fillMaxWidth(),
+        backgroundColor = if (enabled) {
+            QuackColor.White
+        } else {
+            QuackColor.Gray4
+        },
+        border = if (enabled) {
+            QuackBorder(
                 width = 1.dp,
-                brush = if (enabled) {
-                    QuackColor.DuckieOrange.toBrush()
-                } else {
-                    QuackColor.Unspecified.toBrush()
-                },
-                shape = RoundedCornerShape(size = 8.dp),
+                color = QuackColor.DuckieOrange,
             )
-            .padding(
+        } else {
+            null
+        },
+        shape = RoundedCornerShape(size = 8.dp),
+        onClick = onClick,
+    ) {
+        QuackText(
+            modifier = Modifier.padding(
                 vertical = 8.dp,
                 horizontal = 12.dp,
             ),
-        text = text,
-        typography = QuackTypography.Body1.change(
-            color = if (enabled) {
-                QuackColor.DuckieOrange
-            } else {
-                plainTextColor
-            },
-            textAlign = TextAlign.Center,
-        ),
-        singleLine = true,
-    )
+            text = text,
+            style = QuackTextStyle.Body1.change(
+                color = if (enabled) {
+                    QuackColor.DuckieOrange
+                } else {
+                    plainTextColor
+                },
+                textAlign = TextAlign.Center,
+            ),
+            singleLine = true,
+        )
+    }
 }
